% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{dd_dem}
\alias{dd_dem}
\title{Discrete Choice Predictions (HMNL)}
\usage{
dd_dem(dd, est, cores)
}
\arguments{
\item{dd}{data}

\item{est}{est}

\item{cores}{cores}
}
\value{
Draws of expected choice
}
\description{
Discrete Choice Predictions (HMNL)
}
\seealso{
\code{\link[=dd_est_hmnl]{dd_est_hmnl()}} to generate demand predictions based on this model
}
