% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{dd_dem_sr}
\alias{dd_dem_sr}
\title{Discrete Choice Predictions (HMNL with attribute-based screening)}
\usage{
dd_dem_sr(dd, est, cores)
}
\arguments{
\item{dd}{data}

\item{est}{est}

\item{cores}{cores}
}
\value{
Draws of expected choice
}
\description{
Discrete Choice Predictions (HMNL with attribute-based screening)
}
\seealso{
[dd_est_hmnl_screen()] to generate demand predictions based on this model
}
