% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{dd_dem_srpr}
\alias{dd_dem_srpr}
\title{Discrete Choice Predictions (HMNL with attribute-based screening w/price)}
\usage{
dd_dem_srpr(dd, est, cores)
}
\arguments{
\item{dd}{data}

\item{est}{est}

\item{cores}{cores}
}
\value{
Draws of expected choice
}
\description{
Discrete Choice Predictions (HMNL with attribute-based screening w/price)
}
\seealso{
\code{\link[=dd_est_hmnl_screen]{dd_est_hmnl_screen()}} to generate demand predictions based on this model
}
