% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{dd_est_hmnl_screenpr}
\alias{dd_est_hmnl_screenpr}
\title{Estimate discrete choice model (HMNL, attribute-based screening (including price))}
\usage{
dd_est_hmnl_screen(dd, R=100000, keep=10)
}
\arguments{
\item{dd}{discrete choice data (long format)}

\item{R}{draws}

\item{keep}{thinning}

\item{cores}{no of CPU cores to use (default: auto-detect)}

\item{control}{list containing additional settings}
}
\value{
est ec-draw object (List)
}
\description{
Estimate discrete choice model (HMNL, attribute-based screening (including price))
}
\seealso{
\code{\link[=dd_dem_sr]{dd_dem_sr()}} to generate demand predictions based on this model
}
