% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{dd_prob}
\alias{dd_prob}
\title{Discrete Choice Probabilities (HMNL)}
\usage{
dd_prob(dd, est, cores = NULL)
}
\arguments{
\item{dd}{data}

\item{est}{est}

\item{cores}{cores}
}
\value{
Draws of choice probabilities
}
\description{
Discrete Choice Probabilities (HMNL)
}
\seealso{
\code{\link[=dd_est_hmnl]{dd_est_hmnl()}} to generate demand predictions based on this model
}
