% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{dd_prob_srpr}
\alias{dd_prob_srpr}
\title{Discrete Choice Probabilities (HMNL with attribute-based screening w/ price)}
\usage{
dd_prob_srpr(dd, est, cores = NULL)
}
\arguments{
\item{dd}{data}

\item{est}{est}

\item{cores}{cores}
}
\value{
Draws of expected choice
}
\description{
Discrete Choice Probabilities (HMNL with attribute-based screening w/ price)
}
\seealso{
\code{\link[=dd_est_hmnl_screen]{dd_est_hmnl_screen()}} to generate demand predictions based on this model
}
