% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{dropbox_project_sync_off}
\alias{dropbox_project_sync_off}
\title{Turn Dropbox project sync off to avoid issues with .Rproj.user folders (Windows)}
\usage{
dropbox_project_sync_off()
}
\description{
See: https://community.rstudio.com/t/dropbox-conflicts-with-rproj-user/54059/2
}
