% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{dummify}
\alias{dummify}
\title{Dummy-code select variables in a tibble}
\usage{
dummify(dat, sel)
}
\arguments{
\item{dat}{tibble}

\item{sel}{vector of strings, variables to be replaced by dummy coding}
}
\value{
tibble with dummy variables
}
\description{
Dummy-code select variables in a tibble
}
