% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{dummyvar}
\alias{dummyvar}
\title{Dummy-code a categorical variable}
\usage{
dummyvar(data)
}
\arguments{
\item{data}{one column of categorical data to be dummy-coded}
}
\value{
tibble with dummy variables
}
\description{
Dummy-code a categorical variable
}
\examples{
mytest=data.frame(attribute=factor(c('a','a','b','c','c')))
dummyvar(mytest)
}
