% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_boxplot_MU}
\alias{ec_boxplot_MU}
\title{Generate MU_theta boxplot}
\usage{
ec_boxplot_MU(draws)
}
\arguments{
\item{draws}{ec-draws, output from any ec-model}
}
\description{
Generate MU_theta boxplot
}
\seealso{
\code{\link[=ec_trace_MU]{ec_trace_MU()}} to obtain traceplot
}
