% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_boxplot_screen}
\alias{ec_boxplot_screen}
\title{Generate Screening probability boxplot}
\usage{
ec_boxplot_screen(draws)
}
\arguments{
\item{draws}{ec-draws, output from any ec-model with screening}
}
\value{
draws
}
\description{
Generate Screening probability boxplot
}
\seealso{
\code{\link[=ec_draws_MU]{ec_draws_MU()}} to obtain MU_theta draws,
\code{\link[=ec_trace_screen]{ec_trace_screen()}} to generate traceplot
}
