% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_dem_aggregate}
\alias{ec_dem_aggregate}
\title{Aggregate posterior draws of demand}
\usage{
ec_dem_aggregate(de,groupby)
}
\arguments{
\item{de}{demand draws}

\item{groupby}{groupby grouping variables (as (vector of) string(s))}
}
\value{
Aggregated demand predictions
}
\description{
Aggregate demand draws, e.g. from individual-choice occasion-alternative level to individual level.
(using the new demand draw format)
}
