% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_dem_eval}
\alias{ec_dem_eval}
\title{Evaluate (hold-out) demand predictions}
\usage{
ec_dem_eval(de, true_dem)
}
\arguments{
\item{de}{demand draws (output from vd_dem_x function)}
}
\value{
Predictive fit statistics (MAE, MSE, RAE, bias, hit-probability)
}
\description{
This function obtains proper posterior fit statistics.
It computes the difference between true demand and each draw from the demand posterior.
Then, fit statistics are obtained.
}
