% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_dem_eval_pm}
\alias{ec_dem_eval_pm}
\title{Evaluate (hold-out) demand predictions based on posterior mean predictions}
\usage{
ec_dem_eval_pm(de)
}
\arguments{
\item{de}{demand draws (output from vd_dem_x function)}
}
\value{
Predictive fit statistics (MAE, MSE, RAE, bias, hit-probability)
}
\description{
This function obtains proper posterior fit statistics.
}
