% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_dem_summarise}
\alias{ec_dem_summarise}
\title{Summarize posterior draws of demand}
\usage{
ec_dem_summarise(de,quantiles)
}
\arguments{
\item{de}{demand draws}

\item{quantiles}{Quantiles for Credibility Intervals (default: 90\% interval)}
}
\value{
Summary of demand predictions
}
\description{
Adds summaries of posterior draws of demand to tibble.
(using the new demand draw format)
}
