% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_draws_MU}
\alias{ec_draws_MU}
\title{Obtain MU_theta draws}
\usage{
ec_draws_MU(draws)
}
\arguments{
\item{draws}{ec-draws, output from any ec-model}
}
\value{
draws
}
\description{
Obtain MU_theta draws
}
\seealso{
[ec_draws_screen()] to obtain screening parameter draws (where applicable),
[ec_trace_MU()] to generate a traceplot of MU_theta draws
}
