% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_draws_screen}
\alias{ec_draws_screen}
\title{Obtain Screening probability draws}
\usage{
ec_draws_screen(draws)
}
\arguments{
\item{draws}{ec-draws, output from any ec-model with screening}
}
\value{
draws
}
\description{
Obtain Screening probability draws
}
\seealso{
\code{\link[=ec_draws_MU]{ec_draws_MU()}} to obtain MU_theta draws,
\code{\link[=ec_trace_screen]{ec_trace_screen()}} to generate a traceplot of screening draws
}
