% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_estimates_MU}
\alias{ec_estimates_MU}
\title{Obtain upper level model estimates}
\usage{
ec_estimates_MU(est, quantiles=c(.05,.95))
}
\arguments{
\item{est}{is an echoice draw object (list)}

\item{quantiles}{quantile for CI}
}
\value{
tibble with MU (upper level) summaries
}
\description{
Obtain upper level model estimates
}
\examples{
\dontrun{
est \%>\% ec_estimates_MU
est \%>\% ec_estimates_MU \%>\% 
ggplot(aes(x=par,y=mean))+geom_bar(stat='identity')+
facet_wrap(~attribute,scales = 'free') + coord_flip()
}
}
