% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_estimates_SIGMA}
\alias{ec_estimates_SIGMA}
\title{Obtain posterior mean estimates of upper level covariance}
\usage{
ec_estimates_SIGMA(est)
}
\arguments{
\item{est}{is an echoice draw object (list)}
}
\value{
estimates of upper level covariance
}
\description{
Obtain posterior mean estimates of upper level covariance
}
