% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_gen_err_normal}
\alias{ec_gen_err_normal}
\title{Simulate error realization from Normal distribution}
\usage{
ec_gen_err_normal(ec_dem, seed)
}
\arguments{
\item{ec_dem}{discrete or volumetric choice data, with or without x}

\item{seed}{seed for reproducible error realisations; seet is automatically reset of running this function}
}
\value{
error realizations
}
\description{
Simulate error realization from Normal distribution
}
