% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_lmd_NR}
\alias{ec_lmd_NR}
\title{Log Marginal Density (Newton-Raftery)}
\usage{
ec_lmd_NR(M)
}
\arguments{
\item{M}{echoice draw object}
}
\value{
tibble with LMDs (first 25\% of draws, next 25\% of draws, ...)
}
\description{
Very rough approximation of LMD based on Newton-Raftery.
It is not the most accurate, but a very fast method.
Draws are split in 4 equal parts from start to finish, and LMD
is computed for each part. This helps to double-check convergence.
}
