% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_screenprprob_sr}
\alias{ec_screenprprob_sr}
\title{Screening probabilities of choice alternatives (w/ price)}
\usage{
ec_screenprprob_sr(xd, est, cores = NULL)
}
\arguments{
\item{xd}{data}

\item{est}{ec-model draws}

\item{cores}{(optional) cores}
}
\value{
Draws of screening probabilities of choice alternatives
}
\description{
Obtain draws of screening probabilities of choiec alternatives
}
