% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_summarize_attrlvls}
\alias{ec_summarize_attrlvls}
\title{Summarise attributes and levels}
\usage{
ec_summarize_attrlvls(data_in)
}
\arguments{
\item{data_in}{long-format choice data}
}
\value{
tibble with summary
}
\description{
This functions looks for categorical attributes and summaries their levels
This is helpful when evaluating a new choice data file.
}
