% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_trace_MU}
\alias{ec_trace_MU}
\title{Generate MU_theta traceplot}
\usage{
ec_trace_MU(draws)
}
\arguments{
\item{draws}{ec-draws, output from any ec-model}
}
\description{
Generate MU_theta traceplot
}
\seealso{
\code{\link[=ec_boxplot_MU]{ec_boxplot_MU()}} to obtain boxplot
}
