% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_trace_screen}
\alias{ec_trace_screen}
\title{Generate Screening probability traceplot}
\usage{
ec_trace_screen(draws)
}
\arguments{
\item{draws}{ec-draws, output from any ec-model with screening}
}
\value{
draws
}
\description{
Generate Screening probability traceplot
}
\seealso{
[ec_draws_MU()] to obtain MU_theta draws,
[ec_boxplot_screen()] to generate boxplot
}
