% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_undummy_lowhigh}
\alias{ec_undummy_lowhigh}
\title{Convert dummy-coded variables to low/high factor}
\usage{
ec_undummy_lowhigh(vec_in)
}
\arguments{
\item{vec_in}{A vector of dummy-coded variables (0/1)}
}
\value{
A factor vector with levels "low" and "high"
}
\description{
Convert dummy-coded variables to low/high factor
}
\examples{
ec_undummy_lowhigh(c(0,1,0,1,1))


}
