% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_undummy_lowmediumhigh}
\alias{ec_undummy_lowmediumhigh}
\title{Convert dummy-coded variables to low/medium/high factor}
\usage{
ec_undummy_lowmediumhigh(vec_in)
}
\arguments{
\item{vec_in}{A vector of dummy-coded variables (0/1/2)}
}
\value{
A factor vector with levels "low", "medium" and "high"
}
\description{
Convert dummy-coded variables to low/medium/high factor
}
\examples{
ec_undummy_lowmediumhigh(c(0,1,2,1,0,2))


}
