% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_undummy_yesno}
\alias{ec_undummy_yesno}
\title{Convert dummy-coded variables to yes/no factor}
\usage{
ec_undummy_yesno(vec_in)
}
\arguments{
\item{vec_in}{A vector of dummy-coded variables (0/1)}
}
\value{
A factor vector with levels "no" and "yes"
}
\description{
Convert dummy-coded variables to yes/no factor
}
\examples{
ec_undummy_yesno(c(0,1,0,1,1))


}
