\name{echoice2-package}
\alias{echoice2-package}
\alias{echoice2}
\docType{package}
\title{\packageTitle{echoice2}}
\description{\packageDescription{echoice2}}
\section{Package Content}{\packageIndices{echoice2}}
\author{\packageAuthor{echoice2}}
\section{Maintainer}{\packageMaintainer{echoice2}}
\keyword{package}
