% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{get_attr_lvl}
\alias{get_attr_lvl}
\title{Obtain attributes and levels from tidy choice data}
\usage{
get_attr_lvl(tdc)
}
\arguments{
\item{tdc}{tibble}
}
\value{
tibble
}
\description{
Obtain attributes and levels from tidy choice data
}
