% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{\%.\%}
\alias{\%.\%}
\title{Get the attribute of an object}
\usage{
obj \%.\% attrname
}
\arguments{
\item{obj}{The object to get the attribute from.}

\item{attrname}{The name of the attribute to get.}
}
\value{
The attribute of the object.
}
\description{
Get the attribute of an object
}
\examples{
obj <- list(a = 1, b = 2)
attributes(obj)$test="hello"
`\%.\%`(obj, "test")

}
