% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ht_modelMuCompare}
\alias{ht_modelMuCompare}
\title{Generate huxtable for model comparison}
\usage{
ht_modelMuCompare(model_list)
}
\description{
This requires huxtable package
Pass a list of dd or vd models in a named list to this function, and it will generate a huxtable.
This will include posterior mean and standard deviations of the upper-level parameters
The table can easily by converted to latex using huxtable functions.
}
