% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ht_screenCompare}
\alias{ht_screenCompare}
\title{Generate huxtable for model comparison - screening vs no screening}
\usage{
ht_screenCompare(vdm, vdmsrpr)
}
\arguments{
\item{vdm}{vdm object}

\item{vdmsrpr}{vdmsrpr object}
}
\description{
This requires huxtable package
Pass a list of dd or vd models in a named list to this function, and it will generate a huxtable.
This will include posterior mean and standard deviations of the upper-level parameters
The table can easily by converted to latex using huxtable functions.
}
