% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{logMargDenNR}
\alias{logMargDenNR}
\title{Log Marginal Density (Newton-Raftery)}
\usage{
logMargDenNR(ll)
}
\arguments{
\item{ll}{vector of likelihood draws}
}
\value{
LMD (double)
}
\description{
Log Marginal Density (Newton-Raftery)
}
\examples{
\dontrun{
logMargDenNR(ll)dd_dem_sr
}
}
