% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{logMargDenNRu}
\alias{logMargDenNRu}
\title{Log Marginal Density (Newton-Raftery)}
\usage{
logMargDenNRu(ll)
}
\arguments{
\item{ll}{A vector of log-likelihood values (i.e., draws)}
}
\value{
A single numeric value representing the log marginal density
}
\description{
This function uses the quick-and-dirty Newton-Raftery approximation for log-marginal-density.
}
\details{
Approximation of LMD based on Newton-Raftery.
It is not the most accurate, but a very fast method.
}
\examples{
logll_values <- c(-4000, -4001, -4002)
logMargDenNRu(logll_values)
}
