% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{lol_split}
\alias{lol_split}
\title{Split list-of-lists choice data in two (training and holdout)}
\usage{
lol_split=function(lol_data, nho=1)
}
\arguments{
\item{lol_data}{is a list of lists (one list per unit) with elements A, X, P}
}
\value{
list of two list-of-lists
}
\description{
Split list-of-lists choice data in two (training and holdout)
}
