% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{prep_newprediction}
\alias{prep_newprediction}
\title{Match factor levels between two datasets}
\usage{
prep_newprediction(data_new, data_old)
}
\arguments{
\item{data_new}{New long-format choice data}

\item{data_old}{Old long-format choice data}
}
\value{
long-format choice data
}
\description{
Makes sure the factor levels in \code{data_new} are aligned with \code{data_old}
This is helpful for demand simulations.
}
\examples{
data(icecream)
prep_newprediction(icecream, icecream)

}
