% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_add_prodid}
\alias{vd_add_prodid}
\title{Add product id to demand draws}
\usage{
vd_add_prodid(de)
}
\arguments{
\item{de}{demand draws}
}
\value{
est
}
\description{
This adds a unique product identifier to  demand draw objects.
}
