% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_dem_aggregate_long}
\alias{vd_dem_aggregate_long}
\title{Aggregating demand/choice predictions}
\usage{
vd_dem_aggregate_long(de,groupby)
}
\arguments{
\item{de}{demand draws}

\item{groupby}{groupby grouping variables, will be passed to \code{dplyr::group_by} via \code{dplyr::vars}}
}
\value{
Aggregated demand predictions
}
\description{
Aggregates demand prediction draws
}
