% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_dem_aggregate_summarise}
\alias{vd_dem_aggregate_summarise}
\title{Summarising demand/choice predictions}
\usage{
vd_dem_aggregate_summarise(de,groupby)
}
\arguments{
\item{de}{demand draws}

\item{groupby}{groupby grouping variables, will be passed to `dplyr::group_by` via `dplyr::vars`}
}
\value{
Summary of demand predictions
}
\description{
Summarising demand/choice predictions
}
