% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_dem_vdm}
\alias{vd_dem_vdm}
\title{Demand Prediction (Volumetric demand, EV1 errors)}
\usage{
vd_dem_vdm(vd, tidy = TRUE, est, epsilon_not = NULL, cores = NULL)
}
\arguments{
\item{vd}{data}

\item{tidy}{apply echoice tidier}

\item{est}{ec-model draws}

\item{epsilon_not}{(optional) error realizations}

\item{cores}{(optional) cores}
}
\value{
Draws of expected demand
}
\description{
Generating demand predictions for volumetric demand model.
Reminder: there is no closed-form solution for demand, thus we need to integrate not only over the posterior distribution of parameters and the error distribution.
The function outputs a tibble containing id, task, alt, p, attributes, draws from the posterior of demand.
Eerror realisations can be pre-supplied to the \code{epsilon_not}. This helps create smooth demand curves or conduct optimization.
}
\seealso{
\code{\link[=prep_newprediction]{prep_newprediction()}} to match \code{vd}'s factor levels,
\code{\link[=ec_gen_err_ev1]{ec_gen_err_ev1()}} for pre-generating error realizations and
\code{\link[=vd_est_vdm]{vd_est_vdm()}} for estimating the corresponding model
}
