% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_dem_vdmsrpr}
\alias{vd_dem_vdmsrpr}
\title{Demand Prediction (Volumetric demand, attribute-based screening including price, Normal)}
\usage{
vd_dem_vdmsrpr(vd, est, epsilon_not=NULL, cores=NULL)
}
\arguments{
\item{vd}{data}

\item{est}{ec-model draws}

\item{epsilon_not}{(optional) error realizations}

\item{cores}{(optional) cores}
}
\value{
Draws of expected demand
}
\description{
Generating demand predictions for volumetric demand model with attribute-based screening including price. 
Reminder: there is no closed-form solution for demand, thus we need to integrate not only over the posterior distribution of parameters and the error distribution.
The function outputs a tibble containing id, task, alt, p, attributes, draws from the posterior of demand.
Eerror realisations can be pre-supplied to the `epsilon_not`. This helps create smooth demand curves or conduct optimization.
}
\seealso{
[prep_newprediction()] to match `vd`'s factor levels,
  [ec_gen_err_normal()] for pre-generating error realizations and
  [vd_est_vdm_screenpr()] for estimating the corresponding model
}
