% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_est_vdm}
\alias{vd_est_vdm}
\title{Estimate volumetric demand model, EV1 errors}
\usage{
vd_est_vdm(vd, R=100000, keep=10)
}
\arguments{
\item{vd}{volumetric demand data (long format)}

\item{R}{draws}

\item{keep}{thinning}

\item{cores}{no of CPU cores to use (default: auto-detect)}

\item{control}{list containing additional settings}
}
\value{
est ec-draw object (List)
}
\description{
Estimate volumetric demand model, EV1 errors
}
\examples{
\dontrun{
data(icecream)
icecream_est <- icecream \%>\% vd_est_vdm(R=50000)
}
}
\seealso{
[vd_dem_vdm()] to generate demand predictions based on this model
}
