% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_est_vdm_ss}
\alias{vd_est_vdm_ss}
\title{Estimate volumetric demand model accounting for set size variation (1st order)}
\usage{
vd_est_vdm_ss(
  vd,
  R = 1e+05,
  keep = 10,
  cores = NULL,
  control = list(include_data = TRUE)
)
}
\arguments{
\item{vd}{volumetric demand data (long format) with set size variation}

\item{R}{draws}

\item{keep}{thinning}

\item{cores}{no of CPU cores to use (default: auto-detect)}

\item{control}{list containing additional settings}
}
\value{
est ec-draw object (List)
}
\description{
For more details on the model: https://dx.doi.org/10.2139/ssrn.3418383
This model REQUIRES variation in choice-set size
}
