% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_est_vdmn}
\alias{vd_est_vdmn}
\title{Estimate volumetric demand model (Normal Error)}
\usage{
vd_est_vdmn(vd, R=100000, keep=10)
}
\arguments{
\item{vd}{volumetric demand data (long format)}

\item{R}{draws}

\item{keep}{thinning}

\item{cores}{no of CPU cores to use (default: auto-detect)}

\item{control}{list containing additional settings}
}
\value{
est ec-draw object (List)
}
\description{
Estimate volumetric demand model (Normal Error)
}
\examples{
\dontrun{
icecream_est <- icecream \%>\% vd_est_vdmn(R=50000)
}
}
