% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_LL_vdm}
\alias{vd_LL_vdm}
\title{Log-Likelihood for compensatory volumetric demand model}
\usage{
vd_LL_vdm(draw, vd, fromdraw=1)
}
\arguments{
\item{draw}{draws}

\item{vd}{vd object}

\item{fromdraw}{from which draw (excl. burnin)}
}
\value{
Log-Likelihood
}
\description{
Log-Likelihood for compensatory volumetric demand model
}
