% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_LL_vdm_screenpr}
\alias{vd_LL_vdm_screenpr}
\title{Log-Likelihood for conjunctive-screening (with price) volumetric demand model}
\usage{
vd_LL_vdm_screenpr(draw, vd, fromdraw = 1)
}
\arguments{
\item{draw}{draws}

\item{vd}{vd object}

\item{fromdraw}{from which draw (excl. burnin)}
}
\value{
Log-Likelihood
}
\description{
Log-Likelihood for conjunctive-screening (with price) volumetric demand model
}
