% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_lol_tidy}
\alias{vd_lol_tidy}
\title{Convert list-of-lists choice format to stacked tidy data}
\usage{
vd_lol_tidy(datalist)
}
\arguments{
\item{datalist}{is a list of lists (one list per unit) with elements A, X, P}
}
\value{
tibble/data.frame in stacked format
Columns of output data.frame:
id; task: ask/choice occasion number; alt: alternative/product number; x: quantity; p: price; attributes
}
\description{
Convert list-of-lists choice format to stacked tidy data
}
