% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_long_tidy}
\alias{vd_long_tidy}
\title{vd_tidy_choice data from long-format choice data}
\usage{
vd_long_tidy(longdata)
}
\arguments{
\item{longdata}{tibble}
}
\value{
tibble
}
\description{
vd_tidy_choice data from long-format choice data
}
