% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_thin_draw}
\alias{vd_thin_draw}
\title{Thin echoice-vd draw objects}
\usage{
vd_thin_draw(M, burnin_perc = 0.5, total_draws = NULL)
}
\arguments{
\item{M}{is an echoice draw object (list)}

\item{burnin_perc}{how much burn-in to remove}

\item{total_draws}{how many draws to keep after thinning}
}
\value{
thinned echoice draw object (list)
}
\description{
Thin echoice-vd draw objects
}
