% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{dd_prob}
\alias{dd_prob}
\title{Discrete Choice Probabilities (HMNL)}
\usage{
dd_dem(dd, est, cores)
}
\arguments{
\item{dd}{data}

\item{est}{est}

\item{cores}{cores}
}
\value{
Draws of choice probabilities
}
\description{
Discrete Choice Probabilities (HMNL)
}
\seealso{
[dd_est_hmnl()] to generate demand predictions based on this model
}
