% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{dummify}
\alias{dummify}
\title{Create dummy variables within a tibble}
\usage{
dummify(dat, sel)
}
\arguments{
\item{dat}{A \code{tibble} with the data.}

\item{sel}{A character vector with the name(s) of the variables to be dummied.}
}
\value{
tibble with dummy variables
}
\description{
Create dummy variables within a tibble
}
\examples{
mytest=data.frame(A=factor(c('a','a','b','c','c')), B=1:5)
dummify(mytest,"A")

}
