% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_draws_MU}
\alias{ec_draws_MU}
\title{Obtain MU_theta draws}
\usage{
ec_draws_MU(draws)
}
\arguments{
\item{draws}{A list, 'echoice2' draws object}
}
\value{
A tibble, long format, draws of MU
}
\description{
Obtain MU_theta draws
}
\examples{
data(icecream)
#run MCMC sampler (use way more than 20 draws for actual use
icecream_est <- icecream \%>\% dplyr::filter(id<50) \%>\% vd_est_vdm(R=20)
ec_draws_MU(icecream_est)

}
\seealso{
\code{\link[=ec_draws_screen]{ec_draws_screen()}} to obtain screening parameter draws (where applicable),
\code{\link[=ec_trace_MU]{ec_trace_MU()}} to generate a traceplot of MU_theta draws
}
