% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_draws_screen}
\alias{ec_draws_screen}
\title{Obtain Screening probability draws}
\usage{
ec_draws_screen(draws)
}
\arguments{
\item{draws}{A list, 'echoice2' draws object}
}
\value{
A tibble, long format, draws of MU
}
\description{
Obtain Screening probability draws
}
\examples{
data(icecream)
#run MCMC sampler (use way more than 20 draws for actual use
icecream_scr_est <- icecream \%>\% dplyr::filter(id<50) \%>\% vd_est_vdm_screen(R=20)
ec_draws_screen(icecream_scr_est)

}
\seealso{
\code{\link[=ec_draws_MU]{ec_draws_MU()}} to obtain MU_theta draws,
\code{\link[=ec_trace_screen]{ec_trace_screen()}} to generate a traceplot of screening draws
}
