% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_estimates_SIGMA_corr}
\alias{ec_estimates_SIGMA_corr}
\title{Obtain posterior mean estimates of upper level correlations}
\usage{
ec_estimates_SIGMA_corr(est)
}
\arguments{
\item{est}{is an 'echoice2' draw object (list)}
}
\value{
estimates of upper level correlations
}
\description{
Obtain posterior mean estimates of upper level correlations
}
\examples{
data(icecream)
#run MCMC sampler (use way more than 20 draws for actual use)
icecream_est <- icecream \%>\% dplyr::filter(id<20) \%>\% vd_est_vdm(R=20, cores=2)
icecream_est \%>\% ec_estimates_SIGMA_corr \%>\% round(2)
}
