% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_screenprob_sr}
\alias{ec_screenprob_sr}
\title{Screening probabilities of choice alternatives}
\usage{
ec_screenprob_sr(xd, est, cores=NULL)
}
\arguments{
\item{xd}{data}

\item{est}{ec-model draws}

\item{cores}{(optional) cores}
}
\value{
Draws of screening probabilities of choice alternatives
}
\description{
Obtain draws of screening probabilities of choiec alternatives
}
\examples{
data(icecream)
icecream_est <- icecream \%>\% filter(id<10) \%>\% vd_est_vdm_screen(R=10,  price_screen=TRUE)
ec_screenprob_sr(icecream \%>\% filter(id<10), icecream_est) 

}
