% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{ec_trace_MU}
\alias{ec_trace_MU}
\title{Generate MU_theta traceplot}
\usage{
ec_trace_MU(draws, burnin = 100)
}
\arguments{
\item{draws}{A list, 'echoice2' draws object}

\item{burnin}{burn-in to remove}
}
\value{
A ggplot2 plot containing traceplots of draws
}
\description{
Generate MU_theta traceplot
}
\examples{
\dontrun{
data(icecream)
#run MCMC sampler (use way more than 20 draws for actual use
icecream_est <- icecream \%>\% dplyr::filter(id<10) \%>\% vd_est_vdm(R=10, cores=2)
ec_trace_MU(icecream_est)
}

}
\seealso{
\code{\link[=ec_boxplot_MU]{ec_boxplot_MU()}} to obtain boxplot
}
