% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2-package.R
\docType{package}
\name{echoice2-package}
\alias{echoice2}
\alias{echoice2-package}
\title{echoice2: Choice Models with Economic Foundation}
\description{
echoice2 lets you use economic models of choice from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ninohardt/echoice2}
  \item \url{http://ninohardt.de/echoice2/}
  \item Report bugs at \url{https://github.com/ninohardt/echoice2/issues}
}

}
\author{
\strong{Maintainer}: Nino Hardt \email{me@ninohardt.com} (\href{https://orcid.org/0000-0001-5215-6216}{ORCID})

}
\keyword{internal}
