% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2-package.R
\docType{data}
\name{icecream}
\alias{icecream}
\title{icecream}
\description{
Volumetric Conjoint data, ice cream category
}
\details{
Data from volumetric conjoint analysis in the ice cream category. 300 respondents total.
Volumetric demand in units of 4 ounces each. Attributes include brand name, flavor, and container size.
}
