% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{icecream_discrete}
\alias{icecream_discrete}
\title{icecream_discrete}
\description{
icecream_discrete
}
\details{
Data from discrete choice conjoint analysis in the ice cream category. 300 respondents total.
Attributes include brand name, flavor, and container size.
}
