% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_LL_vdm_screenpr}
\alias{vd_LL_vdm_screenpr}
\title{Log-Likelihood for conjunctive-screening (with price) volumetric demand model}
\usage{
vd_LL_vdm_screenpr(draw, vd, fromdraw = 1)
}
\arguments{
\item{draw}{A list, 'echoice2' draws object}

\item{vd}{A tibble, tidy choice data (before dummy-coding)}

\item{fromdraw}{An integer, from which draw onwards to compute LL (i.e., excl. burnin)}
}
\value{
N x Draws Matrix of log-Likelihood values
}
\description{
Log-Likelihood for conjunctive-screening (with price) volumetric demand model
}
\examples{
#analogous to vd_LL_vdm

}
