% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_LL_vdmss}
\alias{vd_LL_vdmss}
\title{Log-Likelihood for volumetric demand model with set-size variation}
\usage{
vd_LL_vdmss(draw, vd, fromdraw = 1)
}
\arguments{
\item{draw}{draws}

\item{vd}{vd object}

\item{fromdraw}{from which draw (excl. burnin)}
}
\value{
Log-Likelihood
}
\description{
Log-Likelihood for volumetric demand model with set-size variation
}
