% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_lol_tidy}
\alias{vd_lol_tidy}
\title{Tidies Volumetric Data into Long Format}
\usage{
vd_lol_tidy(datalist)
}
\arguments{
\item{datalist}{A list of dataframes.}
}
\value{
A dataframe in long format.
Columns of output data.frame:
id; task: ask/choice occasion number; alt: alternative/product number; x: quantity; p: price; attributes
}
\description{
Takes a list of dataframes and tidies each element into a long format.
}
