% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_long_tidy}
\alias{vd_long_tidy}
\title{Generate tidy choice data with dummies from long-format choice data}
\usage{
vd_long_tidy(longdata)
}
\arguments{
\item{longdata}{tibble}
}
\value{
tibble
}
\description{
Generate tidy choice data with dummies from long-format choice data
}
\examples{
data(icecream)
vd_long_tidy(icecream)

}
