% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echoice2.R
\name{vd_prepare_nox}
\alias{vd_prepare_nox}
\title{Prepare choice data for analysis (without x being present)}
\usage{
vd_prepare_nox(dt, Af = NULL)
}
\arguments{
\item{dt}{is tidy choice data (columns: id, task, alt, x, p, attributes)}

\item{Af}{Af}
}
\value{
list containing information for estimation functions
}
\description{
This is mostly for internal use
}
